+++ src/documentation_glossary.jl
@@ -159,6 +159,9 @@ define!(
     :Math, :vector_transport, :symbol, (a = "⋅", b = "⋅") -> raw"\mathcal T_{" * "$a←$b" * "}"
 )
 define!(:Math, :vector_transport, :name, "the vector transport")
+define!(
+    :Math, :VT, (a = "⋅", b = "⋅") -> _math(:vector_transport, :symbol, a, b)
+)
 _math(args...; kwargs...) = glossary(:Math, args...; kwargs...)
 
 #